advanced-transformations-service/requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
gunicorn==21.2.0
Pillow==10.0.0
rembg==2.0.50
requests==2.31.0
Werkzeug==2.3.6
scikit-image==0.21.0
scipy==1.11.1
sympy==1.12
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
basicsr>=1.3.3.11
facexlib>=0.2.0.3
gfpgan>=0.2.1
opencv-python
Pillow
torch>=1.7
torchvision
tqdm
realesrgan